Move ViewingParameters.hpp in the Data directory to match its namespace.
